Split SvgColourServer
